# src/Web.Client/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/completed-products/product/{productId}": {
      "get": {
        "tags": [
          "CompletedProduct"
        ],
        "summary": "Список готовых продуктов по ProductID (справочный ID)",
        "operationId": "GetAllProductsOnStorageByProductId",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompletedProductModel"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/steps/{id}": {
      "get": {
        "tags": [
          "Step"
        ],
        "summary": "Получение этапа по Id",
        "operationId": "GetStepById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Step"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Step"
        ],
        "summary": "Удаление этапа",
        "operationId": "Step_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/steps": {
      "get": {
        "tags": [
          "Step"
        ],
        "summary": "Получение всех этапов",
        "operationId": "GetAllSteps",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "x-position": 1
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            },
            "x-position": 2
          },
          {
            "name": "filterExpression",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "filterValues",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "nullable": true
              }
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Step"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Step"
        ],
        "summary": "Создание этапа",
        "operationId": "Step_Create",
        "requestBody": {
          "x-name": "step",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepCreateRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/role/CreateRole": {
      "post": {
        "tags": [
          "Role"
        ],
        "summary": "Создать новую роль",
        "operationId": "Role_Create",
        "requestBody": {
          "x-name": "name",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/role": {
      "get": {
        "tags": [
          "Role"
        ],
        "summary": "Получить все роли",
        "operationId": "GetAllRoles",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/role/{id}": {
      "get": {
        "tags": [
          "Role"
        ],
        "summary": "Получить роль по id",
        "operationId": "GetRoleById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Получение всех продуктов",
        "operationId": "GetAllProducts",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "x-position": 1
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            },
            "x-position": 2
          },
          {
            "name": "filterExpression",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "filterValues",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string",
                "nullable": true
              }
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "Список продуктов",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValueTupleOfIReadOnlyCollectionOfProductResponseAndInteger"
                }
              }
            }
          },
          "204": {
            "description": "Список продуктов"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "Создать продукт",
        "operationId": "CreateProduct",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "summary": "Обновление продукта по id",
        "operationId": "UpdateProduct",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdateRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Получение продукта по id",
        "operationId": "GetProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "summary": "Удаление продукта по id",
        "operationId": "DeleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/accounts": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Регистрация сотрудника (регистрирует только админ)",
        "operationId": "RegisterAsync",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/login": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Логин сотрудника",
        "operationId": "LoginAsync",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CompletedProductModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          },
          "productDevelopmentId": {
            "type": "string",
            "format": "guid"
          },
          "serialNumber": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "DomainEntity": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          }
        }
      },
...
      "UnitOfMeasure": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Piece"
        ],
        "enum": [
          1
        ]
      },
...
      "Step": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DomainEntity"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "techCardId": {
                "type": "string",
                "format": "guid"
              },
              "operationId": {
                "type": "string",
                "format": "guid"
              },
              "description": {
                "type": "string",
                "nullable": true
              },
              "duration": {
                "type": "string",
                "format": "duration"
              },
              "cost": {
                "type": "number",
                "format": "decimal"
              },
              "unitOfMeasure": {
                "$ref": "#/components/schemas/UnitOfMeasure"
              }
            }
          }
        ]
      },
      "StepCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "techCardId": {
            "type": "string",
            "description": "Ссылка на тех.карту\n            ",
            "format": "guid"
          },
          "operationId": {
            "type": "string",
            "description": "Ссылка на операцию (пайка/мойка/сборка и тд)\n            ",
            "format": "guid"
          },
          "description": {
            "type": "string",
            "description": "Описание задачи\n            ",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "description": "Длительность задачи\n            ",
            "format": "duration"
          },
          "cost": {
            "type": "number",
            "description": "Стоимость задачи\n            ",
            "format": "decimal"
          },
          "unitOfMeasure": {
            "description": "Единица измерения\n            ",
            "oneOf": [
              {
                "$ref": "#/components/schemas/UnitOfMeasure"
              }
            ]
          }
        }
      },
      "RoleResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "ValueTupleOfIReadOnlyCollectionOfProductResponseAndInteger": {
        "type": "object",
        "additionalProperties": false
      },
      "ProductResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "format": "guid",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          },
          "cost": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "format": "guid"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedById": {
            "type": "string",
            "format": "guid"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletedById": {
            "type": "string",
            "format": "guid",
            "nullable": true
          }
        }
      },
      "ProductCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ProductUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "RegisterUserRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "patronymic": {
            "type": "string",
            "nullable": true
          },
          "post": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginUserRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      }
    }
  }
}